feat(data_node): add block txs CBOR endpoints
Proxy `/blocks/{hash_or_number}/txs/cbor` and `/blocks/latest/txs/cbor`
to the data node. Both return `[{ tx_hash, cbor }]` and honor the
standard `count`/`page`/`order` pagination parameters.

The existing `blocks/.../txs` route modules gain sibling `cbor` handlers,
so `txs.rs` becomes a module directory (`txs/{root,cbor}.rs`), matching
the layout already used by e.g. `scripts/{script_hash}` and
`metadata/txs/labels/{label}`.
