+++ crates/integration_tests/tests/data/supported_endpoints.json
@@ -35,7 +35,9 @@
   "/blocks/{hash_or_number}/next",
   "/blocks/{hash_or_number}/previous",
   "/blocks/{hash_or_number}/txs",
+  "/blocks/{hash_or_number}/txs/cbor",
   "/blocks/latest/txs",
+  "/blocks/latest/txs/cbor",
   "/blocks/slot/{slot_number}",
   "/txs/{hash}/delegations",
   "/epochs/{number}/parameters",
