+++ crates/platform/src/api/blocks/hash_or_number/txs.rs
@@ -1,19 +1,2 @@
-use crate::blocks::{BlockData, BlocksPath};
-use crate::{api::ApiResult, server::state::AppState};
-use axum::extract::{Path, Query, State};
-use bf_common::pagination::{Pagination, PaginationQuery};
-
-pub async fn route(
-    State(state): State<AppState>,
-    Query(pagination_query): Query<PaginationQuery>,
-    Path(blocks_path): Path<BlocksPath>,
-) -> ApiResult<Vec<String>> {
-    let block_data = BlockData::from_string(blocks_path.hash_or_number)?;
-    let pagination = Pagination::from_query(pagination_query)?;
-    let data_node = state.data_node()?;
-
-    data_node
-        .blocks()
-        .txs(&block_data.hash_or_number, &pagination)
-        .await
-}
+pub mod cbor;
+pub mod root;
