+++ crates/platform/src/api/blocks/hash_or_number/txs/cbor.rs
@@ -1,0 +1,20 @@
+use crate::blocks::{BlockData, BlocksPath};
+use crate::{api::ApiResult, server::state::AppState};
+use axum::extract::{Path, Query, State};
+use bf_api_provider::types::BlocksTxsCborResponse;
+use bf_common::pagination::{Pagination, PaginationQuery};
+
+pub async fn route(
+    State(state): State<AppState>,
+    Query(pagination_query): Query<PaginationQuery>,
+    Path(blocks_path): Path<BlocksPath>,
+) -> ApiResult<BlocksTxsCborResponse> {
+    let block_data = BlockData::from_string(blocks_path.hash_or_number)?;
+    let pagination = Pagination::from_query(pagination_query)?;
+    let data_node = state.data_node()?;
+
+    data_node
+        .blocks()
+        .txs_cbor(&block_data.hash_or_number, &pagination)
+        .await
+}
